feat: Gemini streaming + 全プロバイダー新 swift-llm-client API 追従 (#1)
* feat: Gemini ストリーミング (streamText) + 全プロバイダー新 API 追従

## 主な変更

### Gemini Streaming (新規)
- LLMCloudGemini/Streaming/GeminiClient+Streaming.swift を新規作成
- :streamGenerateContent?alt=sse エンドポイント経由で SSE ストリーミング
- HTTPStreamingClient (既存) を使用し macOS/Linux 両対応
- AnthropicClient.streamText() と同等のインターフェース

### 全プロバイダー: swift-llm-client 新 API への追従
- AgentCapableClient: thinkingMode: ThinkingMode 引数を追加
- LLMMessage.MessageContent.toolResult: タプルが (toolCallId, name, content) の
  3 要素に変更 (旧: (toolCallId, name, content, isError) の 4 要素)
- ToolResultContent enum (.success/.failure) 経由で content/isError を取得
- ToolChoice.none → ToolChoice.disabled にリネーム
- ToolCallableClient.planToolCalls の systemPrompt が String? → SystemPrompt? に
- Schema Adapters: minimum/maximum/exclusiveMinimum/exclusiveMaximum の
  ConstraintValue を .int(...) → .double(...) に修正

### 影響範囲 (修正ファイル)
- LLMCloudAnthropic/{Provider, Extensions}/*.swift
- LLMCloudGemini/{Provider, Extensions, Schema}/*.swift
- LLMCloudOpenAICompatible/{Client, Converter, Schema, Extensions}/*.swift

### LLMDynamicStructured 廃止対応
- generateDynamic 系メソッドを削除 (DynamicStructured 型が swift-llm-client から削除)
- AnthropicClient+Dynamic.swift / GeminiClient+Dynamic.swift は exclude

### Linux 対応
- swift-api-contract / swift-api-client もローカルパス参照に変更 (検証用)

ビルド: swift build → Build complete! (clean) 警告 0 件

* chore: 依存パッケージを github URL + リリース版に固定

---------

+++ Sources/LLMCloudOpenAICompatible/Schema/OpenAISchemaAdapter.swift
@@ -85,25 +85,25 @@ public struct OpenAISchemaAdapter: ProviderSchemaAdapter {
 
         if let minimum = schema.minimum {
             removedConstraints.append(RemovedConstraint(
-                type: .minimum, fieldPath: fieldPath, value: .int(minimum)
+                type: .minimum, fieldPath: fieldPath, value: .double(minimum)
             ))
         }
 
         if let maximum = schema.maximum {
             removedConstraints.append(RemovedConstraint(
-                type: .maximum, fieldPath: fieldPath, value: .int(maximum)
+                type: .maximum, fieldPath: fieldPath, value: .double(maximum)
             ))
         }
 
         if let exclusiveMinimum = schema.exclusiveMinimum {
             removedConstraints.append(RemovedConstraint(
-                type: .exclusiveMinimum, fieldPath: fieldPath, value: .int(exclusiveMinimum)
+                type: .exclusiveMinimum, fieldPath: fieldPath, value: .double(exclusiveMinimum)
             ))
         }
 
         if let exclusiveMaximum = schema.exclusiveMaximum {
             removedConstraints.append(RemovedConstraint(
-                type: .exclusiveMaximum, fieldPath: fieldPath, value: .int(exclusiveMaximum)
+                type: .exclusiveMaximum, fieldPath: fieldPath, value: .double(exclusiveMaximum)
             ))
         }
